feat: detect certain undesirable patterns so we can remove them later (#5320)
This detects two undesirable patterns so that we can remove them later.

The first is the `_time` parameter being in the group key when window is
used. This wasn't really intended and should have been an error, but it
may have accidentally been done and it's not explicitly an error. We
want to detect it happening so we can communicate those scripts need to
be fixed so we can turn it into an error without disruption.

The second is an undesirable interaction with table find. When table
find is used, the plan is immediately created and evaluated. But, when
this interacts with another plan evalutation, the operation will be
duplicated. This can create inconsistencies from how other parts of the
engine work. We've chosen to have this interaction be an error, but we
want to detect if it's happening so that we can make it an error and so
we can fix any scripts that may be using the interaction inadvertantly.