fix(universe): restore termination defaults for holt_winters (#5323)
Recently I discovered that the Holt Winters examples take > 10s to
complete. This was suspicious so I investigated. I discovered that
recent changes meant that the default epsilon and iteration values
were not used. That is now fixed as the original limits are used.
Tests are updated, it can be seen that the SSE increases a tiny bit as
we are not working as hard to find a good fit, but things are running
faster now too.

Additionally the old neldermead code is removed as it is no longer used.