feat: add PartialOrd and Ord to ast::Position (#5330)
In the flux-lsp, we often have need to figure out if an `ast::Position`
is within an `ast::Range`. The spelling that would be best for this is
`(range.start..range.end).contains(position)`. Unfortunately, we can't
do this today because `ast::Position` doesn't implement `PartialOrd`.
This patch corrects that.

I'm not sure if the test is _really_ all that needed, since the
implementation is derived, but I added it because it's functionality
that we don't want to break in the future.