fix: order of operations for unary operator
Issue:

The formatter was formatting expression with unary operator wrong -(a + b) to -a + b

Fix:
Corrected the precedence of the unary operator to fix the order of operation issue.
Flux has 4 unary operator +/-/not/exists, the existing implementation was not treating +/- different than binary +/-
So to correct the precedence I have given higher precedence to the unary +/-
Also kept the precedence of not and exists untouched to satisfy the current behaviour.

fixes: #5298