fix: Report all errors from all modules from the stateful analyzer
We didn't really have a need to report errors from multiple modules (or packages) before since the standard library was the only thing that were more than a single file/package. This fixes addresses that by checking the errors for all packages and returning those in the stateful analyzer (which is the API used in the C FFI).