Support non-uniformly spaced data in regression
Prior to this change, the linear regression model assumed that the input table data was uniformly spaced (the "x" column was just a uniformly increasing integer). This resulted in incorrect "slope" calculations on data that was nonuniformly-spaced. This change modifies the function's behavior to map the input `_time` columns to a time difference in nanoseconds, and use that value for the calculation of the regression, meaning its output now matches other tools such as Excel's or Libreoffice Calc's "trend lines" analysis.