fix(stdlib): panic in chunking outer joins
An outer join that creates large tables (more than 3000 rows) caused
a panic when splitting the resulting table into chunks. This was
due to the stop parameter becoming lower than that start parameter,
attempting to use nonsensical ranges. Tables larger than 2000 rows
also potentially dropped rows.

The fix is to ensure that the stop parameter is set relative to the start.