feat(array): less leaky string array
Change the behviour of the string array back to the old behaviour
where accessing the Value function returns a string that is backed
by the arrow memory buffer. This avoids data allocations to memory
outside of the memory allocator.

The implementation of array.String has been simplified somewhat as
part of the new behaviour.

There are a number of places where correct behviour relies on copies
of the data being made. To avoid having to fix all of these in the
same PR a temporary ValueCopy function has been added to maintain
the old semantics. This is being used everywhere the Value function
was previously, except for cases where the value is obviously
immediately processed, then discarded.