fix: Make fewer copies when building a string array (#5503)
* fix: Make fewer copies when building a string array

"In the last few months a performance degradation has crept in when
processing flux. One possible cause was the changes that were made
to the way string arrays were built, which always built a full array
even if it would ultimately not be needed. This change ensures that
a full arrow array is only created when necessary.

* chore: update comment for review