feat(stdlib): construct column names with StringBuilder
Update the way row keys and column keys are constructed in the pivot
function. They are now built in a StringBuilder. Previously where they
were constructed the string would be rebuilt with every column added to
the key. If the key only consisted of a single column then a reference
to the column contents would be retained, This reference would no be
counted in the reference count, but would be sufficient to keep an
entire record batch from being garbage collected.