build(deps): update to github.com/apache/arrow-go/v18 (#5540)
Update the arrow version used in flux to github.com/apache/arrow-go/v18.
The Go implementation of arrow has been split away from the main arrow
repository, this will make the arrow module more stable with regards
version numbers. Update to the new version so that it is easier to
update the arrow module in the future.

The version of arrow is strictly speaking part of the external interface
so that change will cause a knock-on upgrade to be required for all
other systems that include flux. This is not epected to be too complex
an update.