feat(array): update String array type
Update flux's String array type to either be an arrow Binary array or an
arrow Dictionary with Binary values. This removes the non-arrow
compatible single string variant, instead using a dictionary to provide
the low memory version for repeated values. A dictionary provides a more
general purpose implementation of the same idea to not keep repeating
identical values.

The StringBuilder still swaps to a standard Binary array after a second
unique value of a String is observed, but does not do so just because a
NULL is added to the array. In the future the heuristic could be
changed to provide memory efficient string representations in other
contexts.

Moving to a completely arrow-compatible interface makes the String array
type much less fragile. It is now possible to use the String array in
any context that an arrow Array can be used, and removes the
special-case code previously required to split a String array.