fix: panic in arrow table buffer
An empty arrow.TableBuffer can panic when calling Len() if it is in the
process of being appended to by a table.BufferBuilder. Normally the
TableBuffer maintains the invariant that the size of Columns is the same
as the size if Values so to check one is equivalent to checkint the
other. However when a BufferBuilder is merging buffers with different
schema it will add new null-valued columns to the TableBuffer. This
checks the TableBuffer.Len() between adding the column and adding the
null-valued column.

In the case that the TableBuffer is empty, but the Columns array is a
zero-length slice, rather than nil, a panic occurs.

There are two fixes here. Either of which fix the symptom, but both are
worthwhile changes. TableBuffer.Len() now checks the length of the slice
it is attempting to access before reading it.
BufferBuilder.normalizeSchema() now checks that the Columns slice length
is 0 (rather than the value being nil) in order to use the shortcut
normalize code path.