feat(array): Run-end encoded string arrays (#5550)
* feat(array): Run-end encoded string arrays

It is very common for flux to be using arrays of strings containing a
single repeated value. This is often the case when processing InfluxDB
tag types. Ad support for run-end encoded string arrays to optimize for
this case. Such arrays are quick to access as they consist of a single
run, but use less memory than a full string array, or even a dictionary
which still needs a full length array to hold the dictionary indexes.

* chore: update comments